feat(host)!: replace wasmer-wasmi backend with wasmer-v8
Swaps the pure-Rust wasmi interpreter for the V8-backed wasmer
backend, patching in the synchwire/wasmer fork so we pick up the
iOS V8 target and the Engine::set_flags_from_string plumbing the
fork adds. The fork is not published; wasmer's full workspace is
patched in via `[patch.crates-io]` on the git branch.

Why:
- wasmerio/wasmer#6500 (merged 2026-04-20) removed both WAMR and
  wasmi from the wasmer crate. Staying on the wasmi backend meant
  pinning to an unmaintained wasmer line.
- V8 with `--jitless` gives an interpreter-only execution mode
  that works on iOS (Apple's no-JIT sandbox). That is what makes
  this the only viable non-JIT wasmer backend for the mobile
  story going forward.

Scope of the swap:
- `crates/host`: rename `module/wasmi.rs` to `module/v8.rs`,
  rename the `wasmer-wasmi` cargo feature to `wasmer-v8`, and
  expose `module::v8::set_flags_from_string` so iOS embedders can
  push `--jitless` before the engine is constructed.
- `test-crates/tests`: update feature flags, cfg gates, and the
  metering stub's error message to reference v8. Drop the
  wasmi-specific `short_circuit` ignore (wasmerio/wasmer#6397
  was a wasmi backend bug; V8 has its own error path).
- `scripts`: rename the two wasmer-wasmi helpers to `-v8` and
  drop the --release workaround that existed for
  wasmerio/wasmer#6392 (a wasmi-only debug-assert issue).
- CI: swap wasmer-wasmi for wasmer-v8 in test.yml and mobile.yml.
  Drop wasmer-v8 from test-windows (v8-custom-builds Windows
  tarball ships v8_monolith.lib, which lacks the wasm-c-api
  symbols wasmer links against). Exclude the x86_64 and armv7
  Android targets from wasmer-v8 (fork only builds aarch64).
  The iOS mobile job now installs brew LLVM for llvm-objcopy and
  adds the aarch64-apple-ios target.
- `flake.nix`: the clang/libclang and LIBCLANG_PATH wiring was
  previously annotated as "for the wasmi feature"; retarget the
  comments at v8 since bindgen still runs on the wasm-c-api
  header.

Known outstanding issue: `cargo test -p tests --features
wasmer-v8` crashes with SIGSEGV when running the second test in
the same process. Single tests pass; the crash is in V8's
per-store lifecycle — a wasmer-side concern rather than a
holochain-wasmer one. Tracked as follow-up.
