+++ scripts/bench-wasmer-v8.sh
@@ -1,6 +1,6 @@
 #! /usr/bin/env bash
 
-cargo bench -p tests --no-default-features --features wasmer-wasmi
+cargo bench -p tests --no-default-features --features wasmer-v8
 
 # it's possible to flamegraph the benchmarks like this:
 #
